card-piv: 'ssize_t' is not defined when compiling with Visual Studio
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@5195 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/libopensc/card-piv.c
@@ -807,7 +807,7 @@ static int piv_read_obj_from_file(sc_card_t * card, char * filename,
 	int r;
 	int f = -1;
 	size_t len;
-	ssize_t res;
+	int res;
 	u8 tagbuf[16];
 	size_t rbuflen;
 	const u8 * body;
